requirements.txt
streamlit-authenticator
qrcode[pil]
streamlit_qrcode_scanner
streamlit-authenticator
pandas
qreader
opencv-python
streamlit-camera-input-live